requirements.txt
pytest
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
decorator==4.4.2
deepdiff==5.0.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
ordered-set==4.0.2
pbr==5.4.5
pluggy==0.13.1
ply==3.11
py==1.9.0
PyYAML==5.3.1
requests==2.24.0
robotframework==3.2.1
robotframework-jsonlibrary==0.3.1
robotframework-requests==0.7.0
robotframework-yamllibrary==0.2.8
six==1.15.0
tox==3.0.0
urllib3==1.25.10
virtualenv==20.0.30
zipp==3.1.0
